refactor(ui): move floating-ui hook to @codeimage/ui
